refactor: restructure DashboardPage components and move to features directory
